refactor(server): extract streaming sse writer
